strategy-engine/pyproject.toml
[project]
name = "alpaca-strategy-engine"
version = "1.0.0"
description = "Deterministic strategy engine with backtesting for Alpaca trading"
requires-python = ">=3.10"
dependencies = [
    "alpaca-py>=0.21.0",
    "pytz",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "psycopg2-binary>=2.9.0",
    "ta>=0.11.0",
    "requests>=2.28.0",
    "finnhub-python>=2.4.0",
]

[project.optional-dependencies]
# Install with: uv pip install -e ".[finbert]" for local FinBERT model
# Requires ~2-3GB disk space for torch + model weights
finbert = [
    "transformers>=4.36.0",
    "torch>=2.0.0",
]
